+++ types/affinda.d.ts
@@ -1559,10 +1559,6 @@ export declare interface ComponentsH65QjbSchemasResumesearchdetailPropertiesSkil
     match?: boolean;
 }
 
-export declare interface ComponentsK7P1F5SchemasResumesearchdetailPropertiesOccupationgroupPropertiesValueItemsAllof1 {
-    match?: boolean;
-}
-
 export declare interface ComponentsN9ShogSchemasResumesearchdetailPropertiesLocationPropertiesValueAllof1 {
     match?: boolean;
 }
@@ -3235,10 +3231,8 @@ export declare interface ResumeSearchDetailManagementLevel {
 
 export declare interface ResumeSearchDetailOccupationGroup {
     missing?: number[];
-    value?: ResumeSearchDetailOccupationGroupValueItem[];
+    value?: OccupationGroupSearchResult[];
 }
-
-export declare type ResumeSearchDetailOccupationGroupValueItem = OccupationGroup & ComponentsK7P1F5SchemasResumesearchdetailPropertiesOccupationgroupPropertiesValueItemsAllof1 & {};
 
 export declare interface ResumeSearchDetailSearchExpression {
     missing?: string[];