+++ src/models/index.ts
@@ -875,6 +875,8 @@ export interface JobDescriptionSearchConfig {
   readonly username?: string;
   /** A list of actions to show in the dropdown in the embedded search tool */
   actions?: JobDescriptionSearchConfigActionsItem[];
+  /** Hide the reset/import toolbar. */
+  hideToolbar?: boolean;
 }
 
 export interface JobDescriptionSearchConfigActionsItem {
@@ -1229,6 +1231,8 @@ export interface ResumeSearchConfig {
   readonly username?: string;
   /** A list of actions to show in the dropdown in the embedded search tool */
   actions?: ResumeSearchConfigActionsItem[];
+  /** Hide the reset/import toolbar. */
+  hideToolbar?: boolean;
 }
 
 export interface ResumeSearchConfigActionsItem {